Fixed issue with missing default initializer on items
Also added event to variables

+++ src/OpenRpg.Items/Extensions/ItemTemplateVariablesExtensions.cs
@@ -11,8 +11,8 @@ namespace OpenRpg.Items.Extensions
         public static void Value(this IItemTemplateVariables variables, int value) => variables[ItemTemplateVariableTypes.Value] = value;
         public static int MaxStacks(this IItemTemplateVariables variables) => (int)variables[ItemTemplateVariableTypes.MaxStacks];
         public static void MaxStacks(this IItemTemplateVariables variables, int value) => variables[ItemTemplateVariableTypes.MaxStacks] = value;
-        public static int Weight(this IItemTemplateVariables variables) => (int)variables[ItemTemplateVariableTypes.Weight];
-        public static void Weight(this IItemTemplateVariables variables, int value) => variables[ItemTemplateVariableTypes.Weight] = value;
+        public static float Weight(this IItemTemplateVariables variables) => variables[ItemTemplateVariableTypes.Weight];
+        public static void Weight(this IItemTemplateVariables variables, float value) => variables[ItemTemplateVariableTypes.Weight] = value;
         public static int SlotType(this IItemTemplateVariables variables) => (int)variables[ItemTemplateVariableTypes.SlotType];
         public static void SlotType(this IItemTemplateVariables variables, int value) => variables[ItemTemplateVariableTypes.SlotType] = value;
     }
